alx_travel_app/requirements.txt
amqp==5.3.1
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.2
certifi==2025.6.15
cffi==1.17.1
chapa==0.1.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
django==4.2.21
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.10
exceptiongroup==1.3.0
Faker==35.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
kombu==5.5.3
mysqlclient==2.2.7
packaging==25.0
pillow==10.4.0
prompt-toolkit==3.0.51
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
